+++ maintenance/optimizer/src/main/java/org/apache/gravitino/maintenance/optimizer/api/recommender/JobExecutionContext.java
@@ -1,0 +1,56 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.gravitino.maintenance.optimizer.api.recommender;
+
+import java.util.Map;
+import org.apache.gravitino.NameIdentifier;
+import org.apache.gravitino.annotation.DeveloperApi;
+
+/**
+ * Immutable description of a recommended optimization job. Produced by {@link StrategyHandler} and
+ * consumed by {@link JobSubmitter}.
+ */
+@DeveloperApi
+public interface JobExecutionContext {
+
+  /**
+   * Target table identifier for the job.
+   *
+   * @return fully qualified table identifier (catalog/schema/table)
+   */
+  NameIdentifier nameIdentifier();
+
+  /**
+   * Free-form job configuration, such as engine parameters (e.g., target file size bytes).
+   *
+   * <p>The {@code StrategyHandler} is free to add additional job configuration besides the job
+   * options specified in the strategy.
+   *
+   * @return immutable map of configuration entries
+   */
+  Map<String, String> jobConfig();
+
+  /**
+   * Job template name to resolve the job in the job submitter.
+   *
+   * @return job template name
+   */
+  String jobTemplateName();
+}
